Move PresenceRouter module_api test to test_presence_router
This test was more concerned with the functionality of PresenceRouter, so I moved it
to the PresenceRouter test file instead.
