+++ tests/module_api/test_api.py
@@ -22,11 +22,7 @@ from synapse.rest import admin
 from synapse.rest.client.v1 import login, presence, room
 from synapse.types import create_requester
 
-from tests.events.test_presence_router import (
-    PresenceRouterTestModule,
-    send_presence_update,
-    sync_presence,
-)
+from tests.events.test_presence_router import send_presence_update, sync_presence
 from tests.test_utils.event_injection import inject_member_event
 from tests.unittest import FederatingHomeserverTestCase, override_config
 
@@ -43,6 +39,7 @@ class ModuleApiTestCase(FederatingHomeserverTestCase):
         self.store = homeserver.get_datastore()
         self.module_api = homeserver.get_module_api()
         self.event_creation_handler = homeserver.get_event_creation_handler()
+        self.sync_handler = homeserver.get_sync_handler()
 
     def make_homeserver(self, reactor, clock):
         return self.setup_test_homeserver(
@@ -377,140 +374,3 @@ class ModuleApiTestCase(FederatingHomeserverTestCase):
                         found_update = True
 
         self.assertTrue(found_update)
-
-    @override_config(
-        {
-            "presence": {
-                "presence_router": {
-                    "module": "%s.%s"
-                    % (
-                        PresenceRouterTestModule.__module__,
-                        PresenceRouterTestModule.__name__,
-                    ),
-                    "config": {
-                        "users_who_should_receive_all_presence": [
-                            "@presence_gobbler1:test",
-                            "@presence_gobbler2:test",
-                            "@far_away_person:island",
-                        ]
-                    },
-                }
-            },
-            "send_federation": True,
-        }
-    )
-    def test_send_local_online_presence_to_with_module(self):
-        """Tests that send_local_presence_to_users sends local online presence to a set
-        of specified local and remote users, with a custom PresenceRouter module enabled.
-        """
-        # Create a user who will send presence updates
-        self.other_user_id = self.register_user("other_user", "monkey")
-        self.other_user_tok = self.login("other_user", "monkey")
-
-        # And another two users that will also send out presence updates, as well as receive
-        # theirs and everyone else's
-        self.presence_receiving_user_one_id = self.register_user(
-            "presence_gobbler1", "monkey"
-        )
-        self.presence_receiving_user_one_tok = self.login("presence_gobbler1", "monkey")
-        self.presence_receiving_user_two_id = self.register_user(
-            "presence_gobbler2", "monkey"
-        )
-        self.presence_receiving_user_two_tok = self.login("presence_gobbler2", "monkey")
-
-        # Have all three users send some presence updates
-        send_presence_update(
-            self,
-            self.other_user_id,
-            self.other_user_tok,
-            "online",
-            "I'm online!",
-        )
-        send_presence_update(
-            self,
-            self.presence_receiving_user_one_id,
-            self.presence_receiving_user_one_tok,
-            "online",
-            "I'm also online!",
-        )
-        send_presence_update(
-            self,
-            self.presence_receiving_user_two_id,
-            self.presence_receiving_user_two_tok,
-            "unavailable",
-            "I'm in a meeting!",
-        )
-
-        # Mark each presence-receiving user for receiving all user presence
-        self.get_success(
-            self.module_api.send_local_online_presence_to(
-                [
-                    self.presence_receiving_user_one_id,
-                    self.presence_receiving_user_two_id,
-                ]
-            )
-        )
-
-        # Perform a sync for each user
-
-        # The other user should only receive their own presence
-        presence_updates, _ = sync_presence(self, self.other_user_id)
-        self.assertEqual(len(presence_updates), 1)
-
-        presence_update = presence_updates[0]  # type: UserPresenceState
-        self.assertEqual(presence_update.user_id, self.other_user_id)
-        self.assertEqual(presence_update.state, "online")
-        self.assertEqual(presence_update.status_msg, "I'm online!")
-
-        # Whereas both presence receiving users should receive everyone's presence updates
-        presence_updates, _ = sync_presence(self, self.presence_receiving_user_one_id)
-        self.assertEqual(len(presence_updates), 3)
-        presence_updates, _ = sync_presence(self, self.presence_receiving_user_two_id)
-        self.assertEqual(len(presence_updates), 3)
-
-        # Test that sending to a remote user works
-        remote_user_id = "@far_away_person:island"
-
-        # Note that due to the remote user being in our module's
-        # users_who_should_receive_all_presence config, they would have
-        # received user presence updates already.
-        #
-        # Thus we reset the mock, and try sending all online local user
-        # presence again
-        self.hs.get_federation_transport_client().send_transaction.reset_mock()
-
-        # Broadcast local user online presence
-        self.get_success(
-            self.module_api.send_local_online_presence_to([remote_user_id])
-        )
-
-        # Check that the expected presence updates were sent
-        expected_users = [
-            self.other_user_id,
-            self.presence_receiving_user_one_id,
-            self.presence_receiving_user_two_id,
-        ]
-
-        calls = (
-            self.hs.get_federation_transport_client().send_transaction.call_args_list
-        )
-        for call in calls:
-            federation_transaction = call.args[0]  # type: Transaction
-
-            # Get the sent EDUs in this transaction
-            edus = federation_transaction.get_dict()["edus"]
-
-            for edu in edus:
-                # Make sure we're only checking presence-type EDUs
-                if edu["edu_type"] != EduTypes.Presence:
-                    continue
-
-                # EDUs can contain multiple presence updates
-                for presence_update in edu["content"]["push"]:
-                    # Check for presence updates that contain the user IDs we're after
-                    expected_users.remove(presence_update["user_id"])
-
-                    # Ensure that no offline states are being sent out
-                    self.assertNotEqual(presence_update["presence"], "offline")
-
-        self.assertEqual(len(expected_users), 0)
